requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.53
Pillow>=10.0.0
pytesseract>=0.3.10
